+++ internal/ldclient/flags.go
@@ -6,13 +6,15 @@ import (
 	"net/http"
 	"net/url"
 
-	ldapi "github.com/launchdarkly/api-client-go/v13"
+	ldapi "github.com/launchdarkly/api-client-go/v15"
 	lcr "github.com/launchdarkly/find-code-references-in-pull-request/config"
+	gha "github.com/launchdarkly/find-code-references-in-pull-request/internal/github_actions"
 	"github.com/launchdarkly/find-code-references-in-pull-request/internal/version"
 	"github.com/pkg/errors"
 )
 
 func GetAllFlags(config *lcr.Config) ([]ldapi.FeatureFlag, error) {
+	gha.Debug("Fetching all flags for project")
 	params := url.Values{}
 	params.Add("env", config.LdEnvironment)
 	activeFlags, err := getFlags(config, params)
@@ -24,7 +26,7 @@ func GetAllFlags(config *lcr.Config) ([]ldapi.FeatureFlag, error) {
 	flags = append(flags, activeFlags...)
 
 	if config.IncludeArchivedFlags {
-		params.Add("archived", "true")
+		params.Add("filter", "state:archived")
 		archivedFlags, err := getFlags(config, params)
 		if err != nil {
 			return []ldapi.FeatureFlag{}, err
@@ -32,6 +34,7 @@ func GetAllFlags(config *lcr.Config) ([]ldapi.FeatureFlag, error) {
 		flags = append(flags, archivedFlags...)
 	}
 
+	gha.Debug("Fetched %d flags", len(flags))
 	return flags, nil
 }